pyproject.toml
[project]
name = "open-perplexity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arize-phoenix-otel>=0.7.1",
    "azure-storage-file-datalake>=12.18.1",
    "boto3>=1.36.19",
    "chainlit>=2.2.0",
    "langchain>=0.3.18",
    "langchain-aws>=0.2.12",
    "langgraph>=0.2.71",
    "openinference-instrumentation-langchain>=0.1.31",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "tavily-python>=0.5.1",
]
